# reports/evidence/ab_router_openapi.json
{"openapi":"3.1.0","info":{"title":"AI Ecosystem Intelligence A/B Router","description":"Step 5 router for controlled model variant traffic splitting.","version":"0.5.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/router/summary":{"get":{"summary":"Router Summary","operationId":"router_summary_router_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouterSummary"}}}}}}},"/rag-answer":{"post":{"summary":"Routed Rag Answer","operationId":"routed_rag_answer_rag_answer_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutedRagRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutedRagResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/feedback":{"post":{"summary":"Feedback","operationId":"feedback_feedback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FeedbackRequest":{"properties":{"request_id":{"type":"string","minLength":3,"title":"Request Id"},"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"},"user_id":{"type":"string","title":"User Id","default":"anonymous"},"comment":{"type":"string","title":"Comment","default":""}},"type":"object","required":["request_id","rating"],"title":"FeedbackRequest"},"FeedbackResponse":{"properties":{"status":{"type":"string","title":"Status"},"request_id":{"type":"string","title":"Request Id"},"rating":{"type":"integer","title":"Rating"}},"type":"object","required":["status","request_id","rating"],"title":"FeedbackResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RoutedRagRequest":{"properties":{"question":{"type":"string","minLength":2,"title":"Question"},"top_k":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Top K","default":5},"request_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Id"},"user_id":{"type":"string","title":"User Id","default":"anonymous"}},"type":"object","required":["question"],"title":"RoutedRagRequest"},"RoutedRagResponse":{"properties":{"request_id":{"type":"string","title":"Request Id"},"user_id":{"type":"string","title":"User Id"},"selected_model":{"type":"string","title":"Selected Model"},"answer":{"type":"string","title":"Answer"},"latency_ms":{"type":"number","title":"Latency Ms"},"evidence_count":{"type":"integer","title":"Evidence Count"},"evidence":{"items":{"type":"object"},"type":"array","title":"Evidence"}},"type":"object","required":["request_id","user_id","selected_model","answer","latency_ms","evidence_count","evidence"],"title":"RoutedRagResponse"},"RouterSummary":{"properties":{"status":{"type":"string","title":"Status"},"model_a_weight":{"type":"number","title":"Model A Weight"},"model_b_weight":{"type":"number","title":"Model B Weight"},"log_path":{"type":"string","title":"Log Path"}},"type":"object","required":["status","model_a_weight","model_b_weight","log_path"],"title":"RouterSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}